:heavy_plus_sign: Redux 설정
