+++ package.json
@@ -10,6 +10,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
     "@svgr/webpack": "^8.1.0",
     "@types/node": "20.4.10",
     "@types/react": "18.2.20",
@@ -19,6 +20,7 @@
     "eslint": "8.47.0",
     "eslint-config-next": "13.4.13",
     "next": "13.4.13",
+    "next-redux-wrapper": "^8.1.0",
     "postcss": "8.4.27",
     "react": "18.2.0",
     "react-dom": "18.2.0",
