+++ src/app/layout.tsx
@@ -1,6 +1,8 @@
+// import wrapper from "@/utils/redux/store";
 import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
+import { Provider } from "react-redux";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -9,14 +11,13 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  // const { store, props } = wrapper.useWrappedStore(null);
   return (
+    // <Provider store={store}>
     <html lang="ko-kr">
       <body className={inter.className}>{children}</body>
     </html>
+    // </Provider>
   );
 }
