+++ src/app/page.tsx
@@ -1,4 +1,5 @@
 import Link from "next/link";
+import { Provider } from "react-redux";
 
 export default function App() {
   return (
