+++ src/utils/redux/store.ts
@@ -1,0 +1,6 @@
+import { configureStore } from "@reduxjs/toolkit";
+import modalStateReducer from "./modalState";
+
+export const store = configureStore({
+  reducer: { modalState: modalStateReducer },
+});
